Overhaul
Saving the drawing layer as a transparent png works now. There's the issue of being able to save it from a network drive, if it is not launched from a network drive location.  GUI alignment was slightly fixed and now has icons! However, still needs the ability to dynamically resize.