Restructured + Updated UI
Cleaned up sample folder, and seperated drawer from the logic/navigation class again for cleanliness.
Removed dynamic resizing, as it lead to problems with alignments and just a lot more trouble for an issue the lab doesn't have. It stays at a fixed size, but the size is adjusted according to the user's windowsize.

Can now accomodate color changes.
The outline will change immediately and the fill will update upon drawing the next blob.